refactor(embed): split index/ai/reconcile.rs god-module into cohesive siblings
The 2.5k-line reconcile module (84 functions) fused model-manifest lifecycle,
cached-model recovery, install/activate, status/plan, and the embed reconcile
loop. Move each concern into a sibling under reconcile/ (manifest,
model_lifecycle, status, embed_loop); mod.rs curates the surface. Inline test
modules move next to the code they exercise.

No behavior change: pure move + call-site migration, no re-export shims.
Verified core + mcp + cli all compile (no cross-crate visibility change).
