+++ crates/rag-rat-core/src/index/ai/reconcile/embed_loop.rs
@@ -1,465 +1,9 @@
 use std::collections::HashMap;
 
-use super::*;
+use super::super::*;
 use crate::config::RemoteEmbeddingConfig;
-use crate::embedding_models::{Backend, EMBEDDING_MODELS, HASH_MODEL_ID, spec};
-// `FASTEMBED_MODEL_ID` is only referenced by the fastembed-gated cache recovery (and tests);
-// the manifest's stale-`'v1'` check now scopes to `HASH_MODEL_ID` only (the old fastembed id
-// was renamed to an HF path in #317 and is legacy-cleaned), so the prod import is
-// feature-gated.
-#[cfg(feature = "fastembed")]
-use crate::embedding_models::{FASTEMBED_EMBEDDING_DIM, FASTEMBED_MODEL_ID};
 
 const RECONCILE_SELECT_ID_BATCH_LIMIT: usize = 900;
-
-pub(crate) fn ensure_model_manifest(conn: &Connection) -> anyhow::Result<()> {
-    // Read-first: skip the (write-locking) DML entirely when the manifest already matches what we
-    // would write. `ensure_model_manifest` runs on EVERY `IndexDatabase::open*`, so issuing
-    // unconditional INSERT/UPDATE/DELETE here made every open — including read-only MCP tools —
-    // take the SQLite write lock, serializing them against the watcher and other clients and
-    // surfacing "database is locked" under concurrency (#143). After the first open establishes
-    // the manifest, every later open is a handful of SELECTs and never touches the write lock.
-    if model_manifest_is_current(conn)? {
-        return Ok(());
-    }
-    remove_legacy_models(conn)?;
-    // One row per registered model, straight from the registry — adding a model needs no edit here.
-    // `installed_by_default` is false for EVERY model (including hash): a model is installed only
-    // on explicit `install_model`. `upsert_model` is `ON CONFLICT DO NOTHING`, so this only
-    // seeds rows.
-    for s in EMBEDDING_MODELS {
-        upsert_model(conn, s.model_id, "embedding", Some(s.dim), s.backend.runtime(), false)?;
-    }
-    normalize_embedding_model_versions(conn)?;
-    Ok(())
-}
-
-/// Read-only test of whether `ensure_model_manifest` would be a no-op — i.e. the manifest is
-/// already in its target state. Mirrors exactly the three writes in `ensure_model_manifest`:
-/// no legacy model rows linger, all three current models are present (`upsert_model` is
-/// `ON CONFLICT DO NOTHING`, so presence is the only condition), and no `chunk_embeddings` row
-/// still carries the pre-normalization `'v1'` model_version. Used both to short-circuit the open
-/// write path (#143) and to let the read-only MCP open refuse to serve when a manifest write is
-/// still owed (falling back to the read-write open, which heals once).
-pub(crate) fn model_manifest_is_current(conn: &Connection) -> anyhow::Result<bool> {
-    for model_id in LEGACY_MODEL_IDS {
-        let lingering: bool = conn.query_row(
-            "SELECT EXISTS(SELECT 1 FROM ai_models WHERE model_id = ?1)
-                 OR EXISTS(SELECT 1 FROM chunk_embeddings WHERE model_id = ?1)
-                 OR EXISTS(SELECT 1 FROM index_meta WHERE key = ?2 AND value = ?1)",
-            params![model_id, ACTIVE_EMBEDDING_MODEL_META],
-            |row| row.get(0),
-        )?;
-        if lingering {
-            return Ok(false);
-        }
-    }
-    for s in EMBEDDING_MODELS {
-        let present: bool = conn.query_row(
-            "SELECT EXISTS(SELECT 1 FROM ai_models WHERE model_id = ?1)",
-            params![s.model_id],
-            |row| row.get(0),
-        )?;
-        if !present {
-            return Ok(false);
-        }
-    }
-    // Only `embedding-hash` can still carry the pre-#112 bare `'v1'` model_version: the old
-    // fastembed id was renamed to an HF path in #317 and is now legacy (deleted above), and the new
-    // HF-path id is fresh. Mirror `normalize_embedding_model_versions`.
-    let stale_version: bool = conn.query_row(
-        "SELECT EXISTS(
-             SELECT 1 FROM chunk_embeddings
-             WHERE model_version = 'v1' AND model_id = ?1
-         )",
-        params![HASH_MODEL_ID],
-        |row| row.get(0),
-    )?;
-    Ok(!stale_version)
-}
-
-pub(crate) fn remove_legacy_models(conn: &Connection) -> anyhow::Result<()> {
-    for model_id in LEGACY_MODEL_IDS {
-        conn.execute("DELETE FROM chunk_embeddings WHERE model_id = ?1", params![model_id])?;
-        conn.execute("DELETE FROM ai_models WHERE model_id = ?1", params![model_id])?;
-        // If this legacy id was the ACTIVE model, its active-model meta AND any persisted
-        // remote-config meta (a legacy `ollama-*` id was a remote install — #317) must both go.
-        // Leaving the remote config behind would let `active_embedder` keep reconstructing an
-        // `OpenAiEmbedder` against a now-removed endpoint after the active model fell back to hash,
-        // so clear it whenever we delete the matching active-model meta.
-        let was_active =
-            conn.execute("DELETE FROM index_meta WHERE key = ?1 AND value = ?2", params![
-                ACTIVE_EMBEDDING_MODEL_META,
-                model_id
-            ])?;
-        if was_active > 0 {
-            clear_active_remote_config(conn)?;
-            // ALSO drop the legacy model's freshness-version meta (R3a): otherwise the hash
-            // fallback inherits the removed model's `model_version` key and reports the
-            // wrong freshness. The next install re-stamps it; clearing here keeps the
-            // gap correct.
-            clear_reconcile_meta(conn, ACTIVE_EMBEDDING_MODEL_VERSION_META)?;
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn normalize_embedding_model_versions(conn: &Connection) -> anyhow::Result<()> {
-    // One-time fix for the pre-#112 bare `'v1'` model_version. Only `embedding-hash` is still a
-    // current id: the old `fastembed-all-minilm-l6-v2` was renamed to an HF path in #317 and is now
-    // a LEGACY id (its rows are deleted by `remove_legacy_models`), so it no longer needs
-    // normalizing here.
-    conn.execute(
-        "
-        UPDATE chunk_embeddings
-        SET model_version = 'hash-v1'
-        WHERE model_version = 'v1' AND model_id = 'embedding-hash'
-        ",
-        [],
-    )?;
-    Ok(())
-}
-
-pub(crate) fn recover_cached_fastembed_model(conn: &Connection) -> anyhow::Result<()> {
-    recover_cached_fastembed_model_from(conn, &fastembed_cache_dir())
-}
-
-pub(crate) fn recover_cached_fastembed_model_from(
-    conn: &Connection,
-    cache_dir: &Path,
-) -> anyhow::Result<()> {
-    #[cfg(feature = "fastembed")]
-    {
-        recover_cached_fastembed_model_at(conn, cache_dir)?;
-    }
-    #[cfg(not(feature = "fastembed"))]
-    {
-        let _ = (conn, cache_dir);
-    }
-    Ok(())
-}
-
-#[cfg(feature = "fastembed")]
-pub(crate) fn recover_cached_fastembed_model_at(
-    conn: &Connection,
-    cache_dir: &Path,
-) -> anyhow::Result<()> {
-    if !fastembed_cache_ready(cache_dir) {
-        return Ok(());
-    }
-    let fastembed = model(conn, FASTEMBED_MODEL_ID)?;
-    if !fastembed.installed || fastembed.status != "Ready" {
-        conn.execute(
-            "UPDATE ai_models
-             SET installed = 1, disabled = 0, status = 'Ready', installed_at_ms = ?2,
-                 embedding_dim = ?3, runtime = 'fastembed', last_error = NULL
-             WHERE model_id = ?1",
-            params![
-                FASTEMBED_MODEL_ID,
-                now_ms(),
-                i64::try_from(FASTEMBED_EMBEDDING_DIM).unwrap_or(i64::MAX)
-            ],
-        )?;
-    }
-    if active_embedding_model_is_missing(conn)? {
-        // Activate the recovered model AND stamp its freshness version in ONE call (R3b): a bare
-        // `set_meta(ACTIVE_EMBEDDING_MODEL_META, ...)` without the version would leave
-        // `active_embedding_model_version` returning a STALE legacy key, so new embeddings bake
-        // under the wrong `model_version` and a later install flips it → a spurious full
-        // re-embed.
-        let spec = spec(FASTEMBED_MODEL_ID)
-            .ok_or_else(|| anyhow::anyhow!("unknown model `{FASTEMBED_MODEL_ID}`"))?;
-        activate_model_with_version(conn, FASTEMBED_MODEL_ID, spec.version)?;
-    }
-    Ok(())
-}
-
-#[cfg(feature = "fastembed")]
-pub(crate) fn active_embedding_model_is_missing(conn: &Connection) -> anyhow::Result<bool> {
-    let Some(active_model_id) = meta(conn, ACTIVE_EMBEDDING_MODEL_META)? else {
-        return Ok(true);
-    };
-    let active = conn
-        .query_row(
-            "
-            SELECT model_id, capability, embedding_dim, runtime, installed, disabled, status, \
-             installed_at_ms, last_error
-            FROM ai_models WHERE model_id = ?1
-            ",
-            [active_model_id],
-            model_row,
-        )
-        .optional()?;
-    Ok(match active {
-        Some(active) => validate_ready_model(&active).is_err(),
-        None => true,
-    })
-}
-
-#[cfg(feature = "fastembed")]
-pub(crate) fn fastembed_cache_ready(cache_dir: &Path) -> bool {
-    let repo = cache_dir.join(FASTEMBED_HF_CACHE_REPO_DIR);
-    let Ok(revision) = std::fs::read_to_string(repo.join("refs").join("main")) else {
-        return false;
-    };
-    let revision = revision.trim();
-    !revision.is_empty() && repo.join("snapshots").join(revision).is_dir()
-}
-
-/// Activate `model_id` as the active embedding model AND stamp its freshness `version` in ONE call
-/// — the SINGLE place that writes both metas, so no activation site can set the active model
-/// without its version (the bug R3b fixed in recovery). `install_model` and
-/// `recover_cached_fastembed_model` both go through here.
-pub(crate) fn activate_model_with_version(
-    conn: &Connection,
-    model_id: &str,
-    version: &str,
-) -> anyhow::Result<()> {
-    set_meta(conn, ACTIVE_EMBEDDING_MODEL_META, model_id)?;
-    set_reconcile_meta(conn, ACTIVE_EMBEDDING_MODEL_VERSION_META, version)?;
-    Ok(())
-}
-
-pub(crate) fn install_model(
-    conn: &Connection,
-    model_id: &str,
-    remote: Option<&RemoteEmbeddingConfig>,
-) -> anyhow::Result<ModelInfo> {
-    ensure_model_manifest(conn)?;
-    // The arg is the model_id — the HF path (no aliases, #317). Resolve to its spec and use the
-    // canonical id downstream.
-    let spec =
-        spec(model_id).ok_or_else(|| anyhow::anyhow!("unknown embedding model `{model_id}`"))?;
-    let model_id = spec.model_id;
-    // The PRESENCE of a remote block — NOT `spec.backend` — selects the Ollama transport (#317
-    // rework): any transformer model can be served over Ollama. The SAME ai_models row toggles its
-    // `runtime` local↔ollama. Remote present → reachability + dim probe (no download); else → the
-    // local install for the model's backend.
-    if let Some(remote) = remote {
-        // A remote block serves the model over Ollama, which can only serve TRANSFORMER models. The
-        // CLI passes the config's remote block for WHATEVER model id the user typed, so guard HERE
-        // too — the config-layer guard only checks `config.model`, not an explicit
-        // `models install <other-id>`. Without this, `models install embedding-hash` against a
-        // 384-dim Ollama would mark the hash row `runtime='ollama'` and store the served model's
-        // vectors under the hash id.
-        if spec.backend != Backend::FastEmbed {
-            anyhow::bail!(
-                "remote embedding requires a transformer model, but `{model_id}` is a {} model — \
-                 remove the [llm.embedding.remote] block to install it locally, or select a \
-                 transformer model to serve over the remote backend",
-                spec.backend.runtime()
-            );
-        }
-        install_remote_model(conn, model_id, spec, remote)?;
-    } else {
-        match spec.backend {
-            Backend::Hash => {
-                conn.execute(
-                    "UPDATE ai_models
-                     SET installed = 1, disabled = 0, status = 'Ready', installed_at_ms = ?2,
-                         embedding_dim = ?3, runtime = 'hash', last_error = NULL
-                     WHERE model_id = ?1",
-                    params![model_id, now_ms(), i64::try_from(spec.dim).unwrap_or(i64::MAX)],
-                )?;
-            },
-            Backend::FastEmbed => install_fastembed_model(conn, model_id)?,
-            Backend::Model2Vec => install_model2vec_model(conn, model_id)?,
-            // Transport-only runtime — never a local install target (the remote branch above owns
-            // the Ollama path).
-            Backend::Ollama => anyhow::bail!(
-                "internal error: Backend::Ollama is a transport, not a local install target"
-            ),
-        }
-        // A LOCAL install must drop any remote-config meta a PRIOR Ollama install of this model
-        // left behind — otherwise `active_embedder` reads it back unconditionally and keeps
-        // building an OpenAiEmbedder against the now-removed endpoint instead of the local
-        // model.
-        clear_active_remote_config(conn)?;
-    }
-    // Activate + stamp the freshness version in one call (the single writer). The version must
-    // reflect the runtime actually installed: a remote install uses the endpoint-INDEPENDENT remote
-    // key (so an ephemeral box's new URL with the same `remote.model` does NOT re-embed), and a
-    // local install uses the static `spec.version` (so flipping remote→local re-embeds).
-    let freshness = match remote {
-        Some(remote) => remote_freshness_version(spec, remote),
-        None => spec.version.to_string(),
-    };
-    activate_model_with_version(conn, model_id, &freshness)?;
-    model(conn, model_id)
-}
-
-pub(crate) fn models(conn: &Connection) -> anyhow::Result<Vec<ModelInfo>> {
-    ensure_model_manifest(conn)?;
-    let mut stmt = conn.prepare(
-        "
-        SELECT model_id, capability, embedding_dim, runtime, installed, disabled, status, \
-         installed_at_ms, last_error
-        FROM ai_models
-        ORDER BY capability, model_id
-        ",
-    )?;
-    let rows = stmt.query_map([], model_row)?;
-    collect_rows(rows)
-}
-
-pub(crate) fn status(conn: &Connection) -> anyhow::Result<LlmStatus> {
-    ensure_model_manifest(conn)?;
-    let total_chunks = chunk_count(conn)?;
-    let active_model_id = active_embedding_model_id(conn)?;
-    let embedding = capability_status(conn, "embedding", &active_model_id, total_chunks)?;
-    let fastembed = fastembed_operational_status(conn, &active_model_id, total_chunks)?;
-    let current = embedding.current_artifacts;
-    let stale = embedding.stale_artifacts;
-    let failed = embedding.failed_artifacts;
-    let blocked = embedding.blocked_artifacts;
-    let missing = total_chunks.saturating_sub(current + stale + failed + blocked);
-    let skipped_chunks = fastembed.skipped_embeddings;
-    let eligible_chunks = total_chunks.saturating_sub(skipped_chunks);
-    Ok(LlmStatus {
-        embedding,
-        artifacts: ArtifactCounts {
-            total_chunks,
-            eligible_chunks,
-            skipped_chunks,
-            current,
-            missing,
-            stale,
-            failed,
-            blocked,
-            disabled: 0,
-        },
-        fastembed,
-        last_reconcile: last_reconcile_status(conn)?,
-    })
-}
-
-/// How many chunks in the ACTIVE scope still need embedding (missing / stale / model- or
-/// dim-changed / retryable-failed), using the SAME candidate sizing the reconcile loop uses. The
-/// watcher gates a per-overlay reconcile on `this_changed`; a reconcile that returned `Partial`
-/// (the shared budget ran out mid-pass) leaves a backlog the next pass would skip because the
-/// overlay rows themselves did not change — so the watcher also reconciles an overlay whose `> 0`
-/// pending count means embeddings were left behind (#219 review). Returns 0 when no embedding model
-/// is ready (nothing to retry) rather than erroring — the watcher must not abort a pass over a
-/// missing embedder.
-pub(crate) fn pending_embedding_jobs(conn: &Connection) -> anyhow::Result<u64> {
-    ensure_model_manifest(conn)?;
-    let model_id = active_embedding_model_id(conn)?;
-    let model = model(conn, &model_id)?;
-    if validate_ready_model(&model).is_err() {
-        return Ok(0);
-    }
-    let model_version = active_embedding_model_version(conn, &model_id)?;
-    let dim = usize::try_from(model.embedding_dim.unwrap_or_default()).unwrap_or(0);
-    let scan = EmbeddingScan {
-        model_id: &model_id,
-        model_version: &model_version,
-        dim,
-        max_embedding_chars: DEFAULT_MAX_EMBEDDING_CHARS,
-    };
-    estimated_reconcile_jobs(conn, &scan, &ReconcileOptions::default())
-}
-
-pub(crate) fn reconcile_plan(conn: &Connection) -> anyhow::Result<ReconcilePlan> {
-    ensure_model_manifest(conn)?;
-    let model_id = active_embedding_model_id(conn)?;
-    let model = model(conn, &model_id)?;
-    let model_version = active_embedding_model_version(conn, &model_id)?;
-    let dim = usize::try_from(model.embedding_dim.unwrap_or_default()).unwrap_or(0);
-    let available = validate_ready_model(&model).is_ok();
-    let message = (!available).then(|| model_not_ready_reason(&model));
-    Ok(ReconcilePlan {
-        embeddings: embedding_reconcile_plan(
-            conn,
-            &model,
-            &model_version,
-            dim,
-            available,
-            message,
-        )?,
-        summaries: SummaryReconcilePlan {
-            enabled: false,
-            message: "summaries are not implemented yet".to_string(),
-        },
-    })
-}
-
-pub(crate) fn embedding_reconcile_plan(
-    conn: &Connection,
-    model: &ModelInfo,
-    model_version: &str,
-    dim: usize,
-    available: bool,
-    message: Option<String>,
-) -> anyhow::Result<EmbeddingReconcilePlan> {
-    let jobs = embedding_job_candidates(conn, &model.model_id, model_version, dim, None, false)?;
-    let skipped_by_policy = embedding_policy_skip_summary(conn, DEFAULT_MAX_EMBEDDING_CHARS)?;
-    let mut missing_by_priority = BTreeMap::new();
-    let mut current = 0_u64;
-    let mut missing = 0_u64;
-    let mut stale = 0_u64;
-    let mut model_changed = 0_u64;
-    let mut dim_changed = 0_u64;
-    let mut failed_retryable = 0_u64;
-    let mut failed_waiting = 0_u64;
-    let mut blocked = 0_u64;
-    for job in jobs {
-        let policy = policy_for_job(&job, DEFAULT_MAX_EMBEDDING_CHARS);
-        if !policy.eligible {
-            continue;
-        }
-        let current_artifact = job.embedding_status.as_deref() == Some("Current")
-            && job.source_text_hash.as_deref() == Some(job.text_hash.as_str())
-            && job.model_version.as_deref() == Some(model_version)
-            && job.embedding_dim == Some(i64::try_from(dim).unwrap_or(i64::MAX))
-            && job.embedding_text_version.as_deref() == Some(EMBEDDING_TEXT_VERSION)
-            && job.input_hash.as_deref().is_some_and(|input_hash| {
-                let input = build_embedding_input(&job, DEFAULT_MAX_EMBEDDING_CHARS);
-                input_hash == embedding_input_hash(&model.model_id, model_version, &input.text)
-            });
-        if current_artifact {
-            current += 1;
-            continue;
-        }
-        let reason = job.reason(model_version, dim, now_ms(), DEFAULT_MAX_EMBEDDING_CHARS);
-        match reason {
-            ReconcileReason::Missing => missing += 1,
-            ReconcileReason::SourceChanged => stale += 1,
-            ReconcileReason::InputChanged => stale += 1,
-            ReconcileReason::ModelChanged => model_changed += 1,
-            ReconcileReason::DimChanged => dim_changed += 1,
-            ReconcileReason::RetryAfterFailure => failed_retryable += 1,
-            ReconcileReason::Forced => missing += 1,
-        }
-        *missing_by_priority.entry(priority_label(policy.priority).to_string()).or_default() += 1;
-        if job.embedding_status.as_deref() == Some("Failed")
-            && job.next_retry_after_ms.unwrap_or(0) > now_ms()
-        {
-            failed_waiting += 1;
-        }
-        if job.embedding_status.as_deref() == Some("Blocked") {
-            blocked += 1;
-        }
-    }
-    Ok(EmbeddingReconcilePlan {
-        model_id: model.model_id.clone(),
-        model_version: model_version.to_string(),
-        dim,
-        available,
-        current,
-        missing,
-        stale,
-        model_changed,
-        dim_changed,
-        failed_retryable,
-        failed_waiting,
-        blocked,
-        disabled: u64::from(model.disabled),
-        skipped_total: skipped_by_policy.values().sum(),
-        skipped_by_policy,
-        missing_by_priority,
-        message,
-    })
-}
 
 pub(crate) fn reconcile(
     conn: &Connection,
@@ -1183,51 +727,6 @@ pub(crate) fn finish_reconcile_attempt(
     Ok(())
 }
 
-pub(crate) fn last_reconcile_status(
-    conn: &Connection,
-) -> anyhow::Result<Option<LastReconcileStatus>> {
-    conn.query_row(
-        "
-        SELECT started_at_ms,
-               finished_at_ms,
-               batch_size,
-               processed_chunks,
-               embeddings_written,
-               blocked_chunks,
-               elapsed_ms,
-               input_chars,
-               status,
-               message
-        FROM reconcile_attempts
-        ORDER BY started_at_ms DESC, id DESC
-        LIMIT 1
-        ",
-        [],
-        |row| {
-            let elapsed_ms = u64::try_from(row.get::<_, i64>(6)?).unwrap_or(0);
-            let input_chars = u64::try_from(row.get::<_, i64>(7)?).unwrap_or(0);
-            let embeddings_written = u64::try_from(row.get::<_, i64>(4)?).unwrap_or(0);
-            let elapsed_secs = (elapsed_ms as f64 / 1000.0).max(0.001);
-            Ok(LastReconcileStatus {
-                started_at_ms: row.get(0)?,
-                finished_at_ms: row.get(1)?,
-                batch_size: u64::try_from(row.get::<_, i64>(2)?).unwrap_or(0),
-                processed_chunks: u64::try_from(row.get::<_, i64>(3)?).unwrap_or(0),
-                embeddings_written,
-                blocked_chunks: u64::try_from(row.get::<_, i64>(5)?).unwrap_or(0),
-                elapsed_ms,
-                input_chars,
-                chunks_per_sec: embeddings_written as f64 / elapsed_secs,
-                chars_per_sec: input_chars as f64 / elapsed_secs,
-                status: row.get(8)?,
-                message: row.get(9)?,
-            })
-        },
-    )
-    .optional()
-    .map_err(Into::into)
-}
-
 pub(crate) fn finalize_reconcile_throughput(report: &mut ReconcileReport, elapsed_ms: u128) {
     report.elapsed_ms = u64::try_from(elapsed_ms).unwrap_or(u64::MAX);
     let elapsed_secs = (report.elapsed_ms as f64 / 1000.0).max(0.001);
@@ -1240,115 +739,6 @@ pub(crate) fn finalize_reconcile_throughput(report: &mut ReconcileReport, elapse
     };
 }
 
-pub(crate) fn upsert_model(
-    conn: &Connection,
-    model_id: &str,
-    capability: &str,
-    embedding_dim: Option<usize>,
-    runtime: &str,
-    installed_by_default: bool,
-) -> anyhow::Result<()> {
-    conn.execute(
-        "
-        INSERT INTO ai_models(model_id, capability, embedding_dim, runtime, installed, disabled, \
-         status, installed_at_ms)
-        VALUES (?1, ?2, ?3, ?4, ?5, 0, ?6, ?7)
-        ON CONFLICT(model_id) DO NOTHING
-        ",
-        params![
-            model_id,
-            capability,
-            embedding_dim.map(|dim| i64::try_from(dim).unwrap_or(i64::MAX)),
-            runtime,
-            installed_by_default,
-            if installed_by_default { "Ready" } else { "MissingModel" },
-            installed_by_default.then(now_ms),
-        ],
-    )?;
-    Ok(())
-}
-
-pub(crate) fn install_model2vec_model(conn: &Connection, model_id: &str) -> anyhow::Result<()> {
-    #[cfg(feature = "model2vec")]
-    {
-        let embedder = Model2VecEmbedder::new()?;
-        conn.execute(
-            "UPDATE ai_models
-             SET installed = 1, disabled = 0, status = 'Ready', installed_at_ms = ?2,
-                 embedding_dim = ?3, runtime = 'model2vec', last_error = NULL
-             WHERE model_id = ?1",
-            params![model_id, now_ms(), i64::try_from(embedder.dim()).unwrap_or(i64::MAX)],
-        )?;
-        Ok(())
-    }
-    #[cfg(not(feature = "model2vec"))]
-    {
-        conn.execute(
-            "UPDATE ai_models
-             SET installed = 0, disabled = 0, status = 'MissingRuntime', last_error = ?2
-             WHERE model_id = ?1",
-            params![model_id, MODEL2VEC_MISSING_FEATURE_MESSAGE],
-        )?;
-        anyhow::bail!("{}", MODEL2VEC_MISSING_FEATURE_MESSAGE)
-    }
-}
-
-#[cfg(test)]
-mod manifest_idempotence_tests {
-    use super::*;
-    use crate::storage::IndexConnection;
-
-    // #143: `ensure_model_manifest` runs on every `IndexDatabase::open*`. It must be a no-op (no
-    // write lock) once the manifest is current, or every read tool serializes on the SQLite writer.
-    #[test]
-    fn ensure_model_manifest_does_not_write_when_already_current() {
-        let dir = std::env::temp_dir().join(format!("ragrat-manifest-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        let db = dir.join("index.db");
-
-        // First open establishes the manifest (a write); afterward the read-only check sees it.
-        {
-            let rw = IndexConnection::open(&db).unwrap();
-            crate::index::schema::apply(rw.connection()).unwrap();
-            assert!(
-                !model_manifest_is_current(rw.connection()).unwrap(),
-                "a freshly applied schema has no model rows yet"
-            );
-            ensure_model_manifest(rw.connection()).unwrap();
-            assert!(model_manifest_is_current(rw.connection()).unwrap());
-        }
-
-        // A current manifest means ensure_model_manifest issues NO DML — prove it by running it on
-        // a read-only connection, which would error if any INSERT/UPDATE/DELETE executed.
-        {
-            let ro = IndexConnection::open_read_only_blocking(&db).unwrap();
-            assert!(model_manifest_is_current(ro.connection()).unwrap());
-            ensure_model_manifest(ro.connection())
-                .expect("a current manifest must not write on a read-only connection");
-        }
-
-        // A lingering legacy model row flips the check back to "needs work".
-        {
-            let rw = IndexConnection::open(&db).unwrap();
-            rw.connection()
-                .execute(
-                    "INSERT INTO ai_models(model_id, capability, embedding_dim, runtime, \
-                     installed, disabled, status, installed_at_ms) VALUES (?1, 'embedding', 384, \
-                     'hash', 0, 0, 'MissingModel', NULL)",
-                    params![LEGACY_MODEL_IDS[0]],
-                )
-                .unwrap();
-            assert!(
-                !model_manifest_is_current(rw.connection()).unwrap(),
-                "a lingering legacy model must require a manifest write"
-            );
-        }
-
-        std::fs::remove_dir_all(&dir).ok();
-    }
-}
-
 #[cfg(test)]
 mod freshness_version_tests {
     use std::io::{Read, Write};
@@ -1359,7 +749,7 @@ mod freshness_version_tests {
     use std::time::{Duration, Instant};
 
     use super::*;
-    use crate::embedding_models::{FASTEMBED_MODEL_ID, HASH_MODEL_ID};
+    use crate::embedding_models::{FASTEMBED_MODEL_ID, HASH_MODEL_ID, spec};
 
     /// One-shot HTTP/1.1 stub replying to the install probe's `/api/embed` with a `dim`-wide
     /// vector.
